📚 : update improvements checklist
what: mark divide coverage item complete

why: reflect new tests for negative and float division

how to test: pre-commit run --all-files && pytest --cov=gabriel

Refs: docs/gabriel/IMPROVEMENTS.md

+++ gabriel/utils.py
@@ -13,7 +13,7 @@ def multiply(a: int, b: int) -> int:
     return a * b
 
 
-def divide(a: int, b: int) -> float:
+def divide(a: float, b: float) -> float:
     """Return the result of ``a`` divided by ``b``."""
     if b == 0:
         raise ZeroDivisionError("Cannot divide by zero.")
